pyproject.toml
[project]
name = "context-engine"
version = "0.1.0"
description = "ContextEngine - Intelligent context assembly for LLMs using dependency graphs and semantic search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-javascript>=0.23.1",
    "networkx>=3.6.1",
    "chromadb>=1.5.2",
    "sentence-transformers>=3.3.1",
    "tiktoken>=0.8.0",
    "anthropic>=0.42.0",
    "openai>=1.59.7",
    "fastapi>=0.135.1",
    "uvicorn>=0.34.0",
    "typer>=0.15.1",
    "rich>=13.9.4",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "torch",
    "tree-sitter-typescript>=0.23.2",
    "watchdog>=6.0.0",
    "google-genai>=1.65.0",
    "mcp>=1.26.0",
]

[project.scripts]
context-engine = "cli.main:app"
context-engine-mcp = "mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["indexer", "retriever", "assembler", "query", "llm", "storage", "api", "cli"]
artifacts = ["models.py", "mcp_server.py"]

[tool.hatch.build.targets.wheel.force-include]
"models.py" = "models.py"
"mcp_server.py" = "mcp_server.py"
